python/requirement.txt
pandas
requests
os
HTTPBasicAuth
mysql-connector-python